improve get-orig-source.sh
For the mplayer package, the support for dynamically looking for
svn:externals was integrated. FFmpeg is now the last package that lacks
this support, and since I like to consider tracking
svn://svn.mplayerhq.hu/ffmpeg/branches/0.5 instead of
svn://svn.mplayerhq.hu/ffmpeg/trunk, it seems pretty convenient to be
able to change that url at a single place.