further improvements to README.upstream-upgrade
the checklist now contains all commands in a copy'n'paster compatible
fashion. This makes it more trivial to test an upstream upgrade in a
staging directory before doing this step in the 'production branch'.

also the order was a bit changed. Now the 'master' branch is finalized
before merging into the unstripped.master branch. This avoids
unnecessary merges.