track the 0.5 release branch
For the next upload to debian, we want to stay at the 0.5 release branch
of ffmpeg. It is not too certain if and how long we will stay on this
one, as this depends on the timeline to the next release and the
(active) maintenance on this branch. But for right now, it is surely not
worse than any other svn snapshot!