be more careful with svn:externals in get-orig-source.sh
the ffmpeg 0.5 release branch used to contain an svn:external to
libswscale. However, the branch was changed to include libswscale rather
than reference it. The problem in the script is that it fetches the
svn:external information from the HEAD revision of the branch rather
than the revision that is being checked out. This patch fixes the issue.