relax the dependencies of the -dev packages a bit
Before the -dev package depended on the exact build version. Now this is
weakened to a version range between the build versions and
${source:Upstream-Version}-99, so that every later version up to the
next upstream version still matches.

Strictly speaking this is only necessary for the -unstripped
variants. If some problems arise because of too weak dependencies, this
decision can be reconsidered.