explicitly disable 'dangerous' encoders on the command line
This commit should be more or less a noop in Debian. A visible change is
that the list of disabled encoders is printed by the 'ffmpeg' binary.

In future, I expect that we don't need to strip the source anymore. In
order to keep the encoders out of the binary package, we need to
explicitly disable them on the configure line. This is done with this
commit.