revert to strict internal libavcodec dependencies
In the following commit, we implemented different internal dependencies
to what is specified in the (also generated) shlibs file without using
an shlibs.local file.

commit a1229f8ace3f5cf129685b26d9ef212637caa3b9
Author: Fabian Greffrath <fabian@debian-unofficial.org>
Date:   Wed Jan 28 15:10:58 2009 +0100

    Use an alternative approach to achieve strict internal dependencies
    by calling dh_makeshlibs twice in debian/rules instead of a
    debian/shlibs.local file.

This patch reinstantiates this behavior. The previous commit did not
clearly indicate that it was the intention to change this behavior.

This reverts parts of commit 3ed1b06fc000290545cdb5f92f0cf39f32e60ce7