move presets back to 'ffmpeg' package. Closes #581748
The problem is more serious at it seems. Leaving the presets in
/usr/share/ffmpeg will cause file conflicts on the next SONAME bump. brr.

This fix is actually more complicated as it seems. I proposed to change
the installation directory to include the soname change. Hower, it did
not receive good reception upstream:

http://comments.gmane.org/gmane.comp.video.ffmpeg.devel/109433

In the end, moving it back to the ffmpeg package seems to be the least
evil here.