clean up NEWS.Debian and related documentation
libav provides a NEWS.Debian file under debian/ but does not install
it currently; rename it to libavcodec-dev.NEWS so apt-listchanges can
automatically install it.  Only installing in a -dev package ensures
sysadmins without the development libraries installed don't have to
worry about API changes that will be irrelevant to them.  Just the
libavcodec-dev package should be enough to make the API changes
visible.

While at it, move APIchanges.gz so it is just installed by the various
development libraries and libav-doc.  The details of the API would be
irrelevant to someone just using programs that use the library.  Maybe
APIchanges should just be in libav-doc, but to be conservative this
keeps it in the development libraries too for now.

Now that NEWS is attached to a particular package and shown on
upgrade, it does not seem like a good place to mention the source
package renaming, which does not require any action in response to the
upgrade.  Move the note to README.Debian, where it is more likely to
help the confused reader.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>