Pass --arch to configure
This helps compilation on systems that run 64bit kernels on 32bit
userland. Upstream considers such configurations as "cross
compiling". While on most architectures the configure script gets it
right nevertheless, on architectures such as mips, it is not able to
find out the 'native' build architecture due to various abis. This
change is supposed ti assist the configuration process.